Modified   Cpp/src/fonts/KS0108.cpp Modified   Cpp/src/fonts/KS0108.hpp Modified   Cpp/src/fonts/KS0108/Georgia15x15.cpp Modified   Cpp/src/main.cpp Modified   Cpp/src/painter/FontDef.hpp Modified   Cpp/src/painter/PainterFont.hpp

+++ Cpp/src/painter/PainterFont.hpp
@@ -10,8 +10,8 @@ protected:
 public:
 	struct Size
 	{
-		uint8_t width;
-		uint8_t height;
+		unsigned width;
+		unsigned height;
 	};
 
 protected:
